exercises/4-1/1-custom-web-application/backend/requirements.txt
Flask
mysql-connector-python
flask-cors